+++ src/Mod/Fem/CMakeLists.txt
@@ -20,11 +20,11 @@ INSTALL(
         convert2TetGen.py
         importCcxDatResults.py
         importCcxFrdResults.py
-	importFenicsXML.py
-	exportFenicsXML.py
-	importFenicsXDMF.py
-	exportFenicsXDMF.py
-	importFenicsMesh.py
+	    readFenicsXML.py
+	    writeFenicsXML.py
+	    readFenicsXDMF.py
+	    writeFenicsXDMF.py
+	    importFenicsMesh.py
         importInpMesh.py
         importToolsFem.py
         importVTKResults.py
